add Damages init Assessment various improvements

+++ firebase/functions/index.js
@@ -3,7 +3,7 @@ const functions = require('firebase-functions');
 admin.initializeApp();
 
 exports.setPermissions = functions.https.onCall(async (data, context) => {
-    if (context.auth.uid === "Thej8pt0ilXGQwLp3vnmqarxpog2" && data.uid !== undefined && data.role !== undefined) {
+    if (context.auth && context.auth.uid === "Thej8pt0ilXGQwLp3vnmqarxpog2" && data.uid !== undefined && data.role !== undefined) {
         await admin.auth().setCustomUserClaims(data.uid, { role: data.role });
         return { status: 200 };
     } else {
@@ -12,7 +12,7 @@ exports.setPermissions = functions.https.onCall(async (data, context) => {
 });
 
 exports.deleteObject = functions.https.onCall(async (data, context) => {
-    if (context.auth.token.role === "admin" && data.oid !== undefined) {
+    if (context.auth && context.auth.token.role === "admin" && data.oid !== undefined) {
         await admin.firestore().collection("objects").doc(data.oid).delete();
         await admin.storage().bucket().deleteFiles({ prefix: `objects/${data.oid}/` });
         return { status: 200 };
@@ -22,9 +22,27 @@ exports.deleteObject = functions.https.onCall(async (data, context) => {
 });
 
 exports.deleteInspection = functions.https.onCall(async (data, context) => {
-    if (data.oid !== undefined && data.iid !== undefined) {
+    if (context.auth && data.oid !== undefined && data.iid !== undefined) {
         await admin.firestore().collection("objects").doc(data.oid).collection('inspections').doc(data.iid).delete();
         await admin.storage().bucket().deleteFiles({ prefix: `objects/${data.oid}/inspections/${data.iid}` });
+        return { status: 200 };
+    } else {
+        return { status: 401 };
+    }
+});
+
+exports.deleteDamage = functions.https.onCall(async (data, context) => {
+    if (context.auth && data.oid !== undefined && data.did !== undefined) {
+        const states = await admin.firestore().collection('objects').doc(data.oid).collection('damages').doc(data.did).collection('states').get();
+        if (!states.empty) {
+            const promises = []
+            for (const state of states.docs) {
+                promises.push(state.ref.delete());
+            }
+            await Promise.all(promises);
+        }
+        await admin.firestore().collection("objects").doc(data.oid).collection('damages').doc(data.did).delete();
+        await admin.storage().bucket().deleteFiles({ prefix: `objects/${data.oid}/damages/${data.did}` });
         return { status: 200 };
     } else {
         return { status: 401 };
